requirements.txt
Django
djangorestframework
mysqlclient
python-dotenv